refactor: add agents for sentiment and outcome analysis
